Supporting excluding hugepages from monitoring
Have used this monitoring a lot on Linux servers.
Whenever using kvm, we dedicate a lot of memory to VMs. Using
hugepages, it won't grow nor shrink, so it's not useful to monitor
this part of the memory.

Added an option to exclude hugepages from the total memory, monitoring
only the memory the host has left.